Task 3 harness + Task 2 configs: dual-GPU DDP sweep
calibrate_power.py: launch every workload via torchrun --nproc_per_node=2 (both
GPUs visible, no CUDA_VISIBLE_DEVICES pin, PYTHONUNBUFFERED=1). PowerSampler sums
both-GPU power for the energy integral and keeps per-GPU util; BytesSampler sums
DRAM-active (field 1005) over both GPUs; NvlinkSampler runs on the same window.
Records tagged ddp/mode=fp16_ddp/precision + per-GPU util + NVLink/PCIe bytes +
n_params, so fp16-DDP records never pool with the legacy fp32 single-GPU fit.
measure_step_rate() auto-sizes steps.

eval_power_monitor.py: CONFIGS replaced with the 91-config fp16-AMP DDP sweep
(families A/B/C/E/F/G/H/I); CONFIG_DEFAULTS -> fp16/nhead8/ff4096; both-GPU >=80%
frontier gate (avg_gpu_pct = min per-GPU); steps auto-sized to ~TARGET_ACTIVE_S;
OOM -> excluded record (not a crash); incremental records dump after every config;
--pool all/fp16/fp32. detect_flops roofline ASSUMED_PRECISION -> FP16 (peak 125).

+++ detect_flops.py
@@ -39,8 +39,10 @@ V100_PROFILE = {
     "PEAK_BW_BYTES_S":  900.0e9,       # HBM2 peak bandwidth
 }
 
-# Roofline is an auxiliary signal. On V100 the matmul default is FP32.
-ASSUMED_PRECISION = "FP32"
+# Roofline is an auxiliary signal. The DDP regime trains in FP16 AMP (that's what
+# the V100 tensor cores are for), so the roofline uses the FP16 tensor-core peak
+# (125 TFLOPS). fp32 remains a valid contrast precision but is no longer the norm.
+ASSUMED_PRECISION = "FP16"
 
 REPO_ROOT = os.path.dirname(os.path.abspath(__file__))
 # /var/log needs root; on this box the daemon runs unprivileged, so the SQLite
